requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
albumentations>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
python-Levenshtein>=0.21.0
python-doctr>=0.7.0
regex>=2023.0.0
